fixing issue with long timeouts
