--- packages/core-node/src/init-environment-server.js
@@ -1,10 +1,0 @@
-#!/usr/bin/env node
-
-const { normalize, join } = require('path');
-
-if (__dirname.includes(normalize('/packages/scripts'))) {
-    require('@ts-tools/node/r');
-} 
-
-require(join(__dirname,'run-environment-server'));
-
