+++ packages/scripts/src/types.ts
@@ -1,5 +1,11 @@
-import { AsyncEnvironment, EnvironmentContext, IComConfig, SomeFeature, TopLevelConfig } from '@wixc3/engine-core';
-import { IEnvironment, IFeatureDefinition } from './analyze-feature';
+import {
+    AsyncEnvironment,
+    EnvironmentContext,
+    EnvironmentTypes,
+    IComConfig,
+    SomeFeature,
+    TopLevelConfig
+} from '@wixc3/engine-core';
 import { IRunNodeEnvironmentsOptions } from './run-socket-server';
 
 export type JSRuntime = 'web' | 'webworker' | 'node';
@@ -168,6 +174,56 @@ export interface RemoteProcess {
     on: (event: 'message', handler: (message: ICommunicationMessage) => unknown) => void;
     postMessage: (message: ICommunicationMessage) => unknown;
     terminate?: () => void;
+}
+
+export interface IFeatureModule {
+    /**
+     * Feature name.
+     * @example "gui" for "gui.feature.ts"
+     */
+    name: string;
+
+    /**
+     * Absolute path pointing to the feature file.
+     */
+    filePath: string;
+
+    /**
+     * Actual evaluated Feature instance exported from the file.
+     */
+    exportedFeature: SomeFeature;
+
+    /**
+     * Exported environments from module.
+     */
+    exportedEnvs: IEnvironment[];
+
+    /**
+     * If module exports any `processingEnv.use('worker')`,
+     * it will be set as `'processing': 'worker'`
+     */
+    usedContexts: Record<string, string>;
+}
+
+export interface IEnvironment {
+    type: EnvironmentTypes;
+    name: string;
+    childEnvName?: string;
+}
+
+export interface IConfigDefinition {
+    name: string;
+    filePath: string;
+    envName?: string;
+}
+
+export interface IFeatureDefinition extends IFeatureModule {
+    contextFilePaths: Record<string, string>;
+    envFilePaths: Record<string, string>;
+    dependencies: string[];
+    scopedName: string;
+    resolvedContexts: Record<string, string>;
+    isRoot: boolean;
 }
 
 export const isEnvironmentStartMessage = (message: ICommunicationMessage): message is IEnvironmaneStartMessage =>
